# [Fix] Cache - Avoiding expensive operations when cache isn't available